Refactor Offers resource to OffersV2 and update related code for compatibility
