+++ src/resources/OffersV2.php
@@ -1,0 +1,42 @@
+<?php
+/**
+ * @author    : Jakiboy
+ * @package   : Amazon Product Advertising API Library (v5)
+ * @version   : 1.3.x
+ * @copyright : (c) 2019 - 2025 Jihad Sinnaour <mail@jihadsinnaour.com>
+ * @link      : https://jakiboy.github.io/apaapi/
+ * @license   : MIT
+ *
+ * This file if a part of Apaapi Lib.
+ */
+
+namespace Apaapi\resources;
+
+use Apaapi\lib\Resource;
+
+/**
+ * Apaapi <OffersV2> : High level resource.
+ * @see https://webservices.amazon.com/paapi5/documentation/offersV2.html
+ */
+final class OffersV2 extends Resource
+{
+	/**
+	 * Set items.
+	 */
+	public function __construct()
+	{
+		$this->items = [
+			'Listings.Availability',
+
+			'Listings.Condition',
+
+			'Listings.IsBuyBoxWinner',
+			'Listings.LoyaltyPoints',
+			'Listings.MerchantInfo',
+
+			'Listings.Price',
+
+			'Listings.Type'
+		];
+	}
+}
